+++ lib/models/text_spoiler_configs.dart
@@ -125,6 +125,7 @@ class TextSpoilerConfig extends SpoilerConfig {
                 speed: particleSpeed ?? this.particleConfig.speed,
                 color: particleColor ?? this.particleConfig.color,
                 maxParticleSize: maxParticleSize ?? this.particleConfig.maxParticleSize,
+                shapePreset: this.particleConfig.shapePreset,
               )
             : this.particleConfig);